refactor: extract datetime logic into reusable modules
